refactor: update x profile
