+++ src/commands/index.ts
@@ -1,8 +1,7 @@
 export { default as getAsteroidsCmd } from './asteroids';
 export { default as getBarrelRollCmd } from './barrelroll';
-export { default as getBlogCmd } from './blog';
 export { default as getCVCmd } from './cv';
 export { default as getGithubCmd } from './github';
 export { default as getLinkedInCmd } from './linkedin';
-export { default as getTwitterCmd } from './twitter';
 export { default as getVersion } from './version';
+export { default as getXCmd } from './x';
