+++ src/components/AsteroidsGame/models/Asteroid.ts
@@ -1,4 +1,4 @@
-// Constants.
+// constants
 import { GameConstants } from '../constants';
 
 // Types.
