+++ src/components/AsteroidsGame/models/Ship.ts
@@ -1,4 +1,4 @@
-// Constants.
+// constants
 import { GameConstants, ShipConstants } from '../constants';
 
 // Types.
