+++ src/components/AsteroidsGame/utils/drawAsteroid.ts
@@ -1,4 +1,4 @@
-// Constants.
+// constants
 import { ShipConstants } from '../constants';
 
 // Models.
