+++ src/components/Button/LinkButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-// Components.
+// components
 import {
   ButtonHoverStyles,
   ButtonOverlay,
@@ -12,7 +12,7 @@ import {
 // Themes.
 import { palette } from '../../theme';
 
-// Utils.
+// utils
 import { getRandomString } from '../../utils';
 
 const buttonOverlayClassName: string = getRandomString(5);
@@ -29,9 +29,9 @@ const StyledLink = styled.a`
   }
 `;
 
-export const LinkButton: React.FC<React.AnchorHTMLAttributes<
-  HTMLAnchorElement
->> = (props: React.AnchorHTMLAttributes<HTMLAnchorElement>) => (
+export const LinkButton: React.FC<
+  React.AnchorHTMLAttributes<HTMLAnchorElement>
+> = (props: React.AnchorHTMLAttributes<HTMLAnchorElement>) => (
   <StyledLink {...props}>
     <ButtonOverlay className={buttonOverlayClassName}>
       <Square />
