+++ src/constants/Links.ts
@@ -1,4 +1,3 @@
-export const BLOG: string = 'https://fromthecornerofakitchentable.blog';
 export const GITHUB: string = 'https://github.com/kieranroneill';
 export const LINKEDIN: string = 'https://www.linkedin.com/in/kieranroneill';
-export const TWITTER: string = 'https://twitter.com/kieranroneill';
+export const X: string = 'https://x.com/magnetartare';
