--- src/descriptions/blog.ts
@@ -1,3 +1,0 @@
-const description: string = 'Opens my blog in a new tab';
-
-export default description;
