+++ src/descriptions/index.ts
@@ -1,8 +1,7 @@
 export { default as asteroidsDescription } from './asteroids';
 export { default as barrelrollDescription } from './barrelroll';
-export { default as blogDescription } from './blog';
 export { default as cvDescription } from './cv';
 export { default as githubDescription } from './github';
 export { default as linkedinDescription } from './linkedin';
-export { default as twitterDescription } from './twitter';
 export { default as versionDescription } from './version';
+export { default as xDescription } from './x';
