--- src/descriptions/twitter.ts
@@ -1,3 +1,0 @@
-const description: string = 'Opens my Twitter profile in a new tab';
-
-export default description;
