+++ src/descriptions/x.ts
@@ -1,0 +1,3 @@
+const description: string = 'Opens my X profile in a new tab';
+
+export default description;
